+++ TODO.md
@@ -10,6 +10,8 @@ $var etc as props
 
 monorepo (maybe using turbo)
 
+// package version check should ignore \* for local workspaces
+
 - ww-library
   - components
   - hooks
