+++ examples/hello-world/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Whimsy Writer</title>
+    <title>Hello World - Whimsy Writer</title>
   </head>
   <body>
     <div class="whimsy-wrapper w-full h-full"></div>
