+++ examples/hello-world/src/__generated/passage-names.d.ts
@@ -1,0 +1,6 @@
+declare global {
+  type Passage = FunctionComponent<unknown>;
+  type PassageName = 'End' | 'Middle' | 'Start' | 'WithMarkdown';
+  type PassageMap = Record<PassageName, Passage>;
+}
+export {};
