+++ examples/hello-world/src/main.tsx
@@ -1,0 +1,11 @@
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import { BaseLayout } from 'whimsy-writer/src/components/base-layout.tsx';
+import './__generated/passage-map.ts';
+import './style/base.css';
+
+createRoot(document.body.firstElementChild!).render(
+  <StrictMode>
+    <BaseLayout />
+  </StrictMode>,
+);
