+++ examples/hello-world/src/passages/middle.tsx
