+++ examples/hello-world/src/save-interface.ts
@@ -1,0 +1,5 @@
+declare global {
+  interface Variables {
+    score: number;
+  }
+}
