+++ examples/hello-world/src/vite-env.d.ts
