+++ package.json
@@ -4,52 +4,14 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "package-version-check": "check-dependency-version-consistency .",
-    "dev": "vite serve -c ./vite/vite.config.ts",
-    "build": "npm run update-passages && tsc -b && vite build -c ./vite/vite.config.ts",
-    "lint": "eslint .",
-    "update-passages": "npx tsx ./vite/update-passages.ts",
-    "watch-passages": "npx tsx ./vite/watch-passages.ts",
-    "__postinstall": "npm run update-passages"
+    "package-version-check": "npm run package-version-check -w repo-utils "
   },
   "workspaces": [
     "packages/*",
     "examples/*",
     "repo-utils"
   ],
-  "dependencies": {
-    "@tailwindcss/typography": "0.5.15",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-icons": "5.3.0",
-    "react-markdown": "9.0.1",
-    "vite-plugin-singlefile": "2.0.2",
-    "zustand": "4.5.5"
-  },
   "devDependencies": {
-    "@babel/parser": "7.25.7",
-    "@babel/traverse": "7.25.7",
-    "@eslint/js": "9.12.0",
-    "@total-typescript/ts-reset": "0.6.1",
-    "@types/babel__traverse": "7.20.6",
-    "@types/node": "22.7.4",
-    "@types/react": "18.3.11",
-    "@types/react-dom": "18.3.0",
-    "@vitejs/plugin-react-swc": "3.7.1",
-    "autoprefixer": "10.4.20",
-    "check-dependency-version-consistency": "4.1.0",
-    "chokidar": "4.0.1",
-    "eslint": "9.12.0",
-    "eslint-plugin-react-hooks": "5.1.0-rc-1460d67c-20241003",
-    "eslint-plugin-react-refresh": "0.4.12",
-    "globals": "15.10.0",
-    "postcss": "8.4.47",
-    "tailwindcss": "3.4.13",
-    "tinyglobby": "0.2.9",
-    "tsx": "4.19.1",
-    "turbo": "2.1.3",
-    "typescript": "5.6.2",
-    "typescript-eslint": "8.8.0",
-    "vite": "5.4.8"
+    "turbo": "2.1.3"
   }
 }
