+++ packages/eslint-config-base/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "eslint-config-base",
+  "devDependencies": {
+    "@eslint/js": "9.12.0",
+    "eslint": "9.12.0",
+    "eslint-plugin-react-hooks": "5.1.0-rc-1460d67c-20241003",
+    "eslint-plugin-react-refresh": "0.4.12",
+    "globals": "15.10.0",
+    "typescript": "5.6.2",
+    "typescript-eslint": "8.8.0"
+  }
+}
