+++ packages/whimsy-writer/package.json
@@ -8,7 +8,8 @@
   "scripts": {
     "build": "vite build",
     "lint": "eslint .",
-    "type-check": "tsc --noEmit"
+    "type-check": "tsc --noEmit",
+    "__postinstall": "npm run update-passages"
   },
   "dependencies": {
     "@babel/parser": "7.25.7",
@@ -20,10 +21,12 @@
     "@eslint/js": "9.12.0",
     "@types/react": "18.3.11",
     "@types/react-dom": "18.3.0",
+    "@types/babel__traverse": "7.20.6",
     "@vitejs/plugin-react-swc": "3.7.1",
     "eslint": "9.12.0",
     "eslint-plugin-react-hooks": "5.1.0-rc-1460d67c-20241003",
     "eslint-plugin-react-refresh": "0.4.12",
+    "eslint-config-base": "*",
     "globals": "15.10.0",
     "typescript": "5.6.2",
     "typescript-eslint": "8.8.0",
