+++ packages/whimsy-writer/src/__backup/vite/utils/template/passage-map-template.ts
