+++ packages/whimsy-writer/src/__backup/vite/utils/write-passage-map.ts
