--- src/base-components/types.ts
@@ -1,3 +1,0 @@
-import { FunctionComponent } from 'react';
-
-export type Passage<T = unknown> = FunctionComponent<T>;
