--- src/save-interface.ts
@@ -1,12 +1,0 @@
-declare global {
-  interface Variables {
-    player: {
-      age: number;
-      inventory: Array<{
-        id: number;
-        qty: number;
-      }>;
-    };
-    score: number;
-  }
-}
